+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.126.0] - 2026-07-18
+
 ### Added
 
 - `gk pr` / `gk issue` / `gk inbox` filter and output flags: